refactor:#2 파일 루트 변경, 버튼 타입 변경
